refactor: rename sqflite.dart to sqflite_service.dart
