+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'screens/mounts_screen/screen.dart';
 import 'screens/rclone_not_installed/screen.dart';
-import 'services/sqflite.dart';
+import 'services/sqflite_service.dart';
 import 'utils/rclone.dart';
 
 void main() async {
